requirements.txt
Django==4.2
djangorestframework==3.15.1
django-debug-toolbar==4.3.0
django-filter==24.2
crispy-bootstrap5==0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
Markdown==3.6
